refactor: renommer components/dashboard en components/modele et deplacer video-task-editor dans components/analyse-video
