+++ src/app/vitrine/page.tsx
@@ -9,11 +9,11 @@
 import { useState, useEffect } from 'react'
 import { motion } from 'framer-motion'
 import { ChevronDown, ChevronUp } from 'lucide-react'
-import TapComponent from '@/components/dashboard/tap-component'
-import GlassComponent from '@/components/dashboard/glass-component'
-import StrawComponent from '@/components/dashboard/straw-component'
-import StormComponent from '@/components/dashboard/storm-component'
-import BubbleComponent from '@/components/dashboard/bubble-component'
+import TapComponent from '@/components/modele/tap-component'
+import GlassComponent from '@/components/modele/glass-component'
+import StrawComponent from '@/components/modele/straw-component'
+import StormComponent from '@/components/modele/storm-component'
+import BubbleComponent from '@/components/modele/bubble-component'
 import { SectionsExplicatives } from './sections-explicatives'
 import ScrollExpandMedia from '@/components/ui/scroll-expansion-hero'
 import FeaturesCards from '@/components/ui/feature-shader-cards'
