+++ src/app/vitrine/sections-explicatives.tsx
@@ -10,11 +10,11 @@
 import { useState } from 'react'
 import { motion } from 'framer-motion'
 import { ChevronDown, ChevronUp } from 'lucide-react'
-import GlassComponent from '@/components/dashboard/glass-component'
-import TapComponent from '@/components/dashboard/tap-component'
-import BubbleComponent from '@/components/dashboard/bubble-component'
-import StormComponent from '@/components/dashboard/storm-component'
-import StrawComponent from '@/components/dashboard/straw-component'
+import GlassComponent from '@/components/modele/glass-component'
+import TapComponent from '@/components/modele/tap-component'
+import BubbleComponent from '@/components/modele/bubble-component'
+import StormComponent from '@/components/modele/storm-component'
+import StrawComponent from '@/components/modele/straw-component'
 import { Slider } from '@/components/ui/slider-number-flow'
 import { InteractiveHoverButton } from '@/components/ui/interactive-hover-button'
