+++ src/components/modele/bubble-component.tsx
