+++ src/components/modele/dashboard-simplified.tsx
