+++ src/components/modele/glass-component.tsx
