+++ src/components/modele/straw-component.tsx
