+++ src/components/modele/tap-component.tsx
